+++ kinto/plugins/quotas/__init__.py
@@ -9,7 +9,13 @@ def includeme(config):
                                           'and collections.',
                               url='https://kinto.readthedocs.io')
 
+    # If StatsD is enabled, monitor execution time of listener.
+    listener = on_resource_changed
+    if config.registry.statsd:
+        key = 'plugins.quotas'
+        listener = config.registry.statsd.timer(key)(on_resource_changed)
+
     # Listen to every resources (except history)
-    config.add_subscriber(on_resource_changed, ResourceChanged,
+    config.add_subscriber(listener, ResourceChanged,
                           for_resources=('bucket', 'group',
                                          'collection', 'record'))